Improving unit test performance, minor refactoring and fixing tests.

+++ src/main/java/bolt/BoltPacket.java
@@ -8,7 +8,7 @@ public interface BoltPacket extends Comparable<BoltPacket> {
 
     void setMessageId(int messageNumber);
 
-    long getDestinationID();
+    int getDestinationID();
 
     void setDestinationID(int destinationID);
